Changes made

+++ Blog/blog_app/views.py
@@ -5,7 +5,8 @@ from blog_app.models import Post,Comment
 from blog_app.forms import PostForm,CommentForm
 from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
 from django.views.generic import (TemplateView,ListView,DetailView,CreateView,
-                                    UpdateView,DeleteView)
+                                    )
+from django.views.generic.edit import UpdateView,DeleteView
 from django.http import Http404
 from django.urls import reverse_lazy
 from django.views import generic
@@ -13,18 +14,36 @@ from django.contrib.auth.mixins import LoginRequiredMixin,PermissionRequiredMixi
 from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth import login,logout,authenticate,get_user_model
 from braces.views import SelectRelatedMixin
+from django.core.exceptions import PermissionDenied
+from django.db.models import Q
+from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 
 User = get_user_model()
 class AboutView(TemplateView):
     template_name = 'about.html'
 
 
-class PostListView(SelectRelatedMixin,ListView):
-    model = Post
-    select_related = ('author')
 
+class PostListView(ListView):
+    model = Post
+    template_name = 'blog_app/post_list.html'
+    paginate_by = 4
     def get_queryset(self):
         return Post.objects.filter(publish_date__lte = timezone.now()).order_by('-publish_date')
+
+class SearchResultView(ListView):
+    model = Post
+    template_name = 'blog_app/search_results.html'
+    paginate_by = 4
+
+    def get_queryset(self, *args, **kwargs):
+        queryset_list = Post.objects.all()
+        query = self.request.GET.get('q')
+        if query:
+            queryset_list = queryset_list.filter(Q(title__icontains=query) | Q(text__icontains=query)).distinct().order_by('-publish_date')
+        if not query:
+            queryset_list = Post.objects.none()
+        return queryset_list
 
 
 class PostDetailView(DetailView):
@@ -42,26 +61,26 @@ class UpdatePostView(LoginRequiredMixin,UpdateView):
     form_class = PostForm
     model = Post
 
+    def get_queryset(self):
+        return Post.objects.filter(author=self.request.user)
+
 class DraftListView(LoginRequiredMixin,ListView):
     login_url = 'accounts/login/'
     redirect_field_name = 'blog_app/post_draft_list.html'
     model = Post
-
-
+    #template_name = 'blog_app/post_draft_list.html'
+    paginate_by = 4
     def get_queryset(self):
-        return Post.objects.filter(publish_date__isnull=True).order_by('-create_date')
+        return Post.objects.filter(author=self.request.user,publish_date__isnull=True).order_by('-create_date')
 
 class DeletePostView(LoginRequiredMixin,DeleteView):
     model = Post
+    def get_queryset(self):
+        return Post.objects.filter(author=self.request.user)
     success_url = reverse_lazy('blog_app:post_list')
 
-
-
-
 ####################################################
 ####################################################
-
-
 
 @login_required
 def post_publish(request,pk):
@@ -84,23 +103,21 @@ def add_comment_to_post(request,pk):
         form = CommentForm()
     return render(request,'blog_app/comment_form.html',{'form':form})
 
-# @login_required
-# def post_remove(request, pk):
-#     post = get_object_or_404(Post, pk=pk)
-#     post.delete()
-#     return redirect('post_list')
-
-
 @login_required
 def comment_approve(request,pk):
     comment = get_object_or_404(Comment,pk=pk)
-    comment.approve()
-    return redirect('blog_app:post_detail',pk=comment.post.pk)
+    item = Post.objects.get(pk=comment.post.pk)
+    if request.user == item.author:
+        comment.approve()
+        return redirect('blog_app:post_detail',pk=comment.post.pk)
+    raise Http404
 
 
 @login_required
 def comment_remove(request,pk):
     comment = get_object_or_404(Comment,pk=pk)
-    post_pk = comment.post.pk
-    comment.delete()
-    return redirect('blog_app:post_detail',pk=post_pk)
+    item = Post.objects.get(pk=comment.post.pk)
+    if request.user == item.author:
+        comment.delete()
+        return redirect('blog_app:post_detail',pk=comment.post.pk)
+    raise Http404
